+++ packages/ios-sdk-ffi/src/token/queries/mod.rs
@@ -1,2 +1,3 @@
 // Token information operations
 pub mod balances;
+pub mod status;
